feat: move search logic to /search
