+++ project/apps/web/app/match/[id]/page.tsx
@@ -4,7 +4,7 @@ import { ArrowLeft } from 'lucide-react';
 import Link from 'next/link';
 import { Suspense } from 'react';
 
-import MatchSkeleton from '@/components/match/MatchSkeleton';
+import MatchSkeleton from '@/components/home/MatchSkeleton';
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar';
 import { Button } from '@/components/ui/button';
 import { useAuthStore } from '@/stores/useAuthStore';
